requirements.txt
django
djangorestframework
djangorestframework_simplejwt
uvicorn
dotenv
whitenoise
mysqlclient
django-cors-headers
pandas
openpyxl